deltaTime 统计的功能有了 ...

+++ Assets/_Scripts/FrameDataDef/HeartbeatCmd.cs
@@ -5,6 +5,6 @@ using UnityEngine;
 public class HeartbeatCmd
 {
     public int Index;
-    public float SendTime;
-    public float RecvTime;
+    public int SendTime; // 时间单位是 ms
+    public int RecvTime;
 }
